refactor agent api
